Move line and bar charts to aggregates and generalise away from sectors
